+++ packages/library-base/src/commonMain/kotlin/io/realm/RealmConfiguration.kt
@@ -133,7 +133,8 @@ public interface RealmConfiguration : Configuration {
                 encryptionKey,
                 deleteRealmIfMigrationNeeded,
                 compactOnLaunchCallback,
-                migration
+                migration,
+                initialDataCallback
             )
         }
     }